S2SRMS: example implementation of the S2SRMS method suggested by R2
Its mainly added for completeness, for feedback/review of other co-authors,
and for R2 to look at it snd its implausible results hould they check this
repository.
We believe that the methods the reviewer suggests (precision estimation on signal
data based on spatial distances of consecutive samples) is not applicable to eye
tracking data collected on professionally produced Hollywood movies. A more
detailed explanation is in the comments of the function, and in our response
letter for the submission.