install and get relevant subdataset data
First reported in #12. At some point during the revision
we must have forgotten to code getting the data into the
function. We likely always had it present locally, and my
previous presentations of the paper were based on a master
branch that did not yet contain the new functions from the
revisions