Make DataLad optional
This change replaces `datalad_get()` with a noop function, whenever
datalad is not available.

This change enables other means of provisioning data, and/or running the
analysis code in environments where datalad is not available.

When datalad is not around, the replacement function produces a message
that reports the data file it would have provisioned.