Correct text label y-position in flowchart
There is some kind of subtle issue with the interpretation of this SVG
(written by Inkscape 0.9.4) by Inkscape 1.2.2, which is used in the
containerized environment created in 2023.

The issue is seemingly confined to the y-coordinate of the "Noise
reduction" text label. The correct label seems to be awefully identical
to the x-coordinate, which makes me suspect a bug in that older Inkscape
version. Unclear, though.

This change manually adjusts the coordinate to visually fit the
published version in the journal.